This branch improved the conformance testing implementation.
The conformance framework is now able to traverse folders to load Querycases from XML files, execute the queries and store the query result to the specified file. The results files are then compared against the expected files using 8 Types of expected outcomes and are written to the specified results file in the same folder. 

The test data and result files are in the /geosparql-benchmarking/conformance_run_data folder. 

CHANGES:
Some classes and methods in the Core-Benchmark and GeoSPARQL-Jena-Benchmark Projects have been altered or added. Please Diff against master. 

TODO:
- The query execution control logic needs to be added to make the benchmarking consistent with the other TestSystems. 
- There is work needed on validating the outcomes and adding more conformance testing implementation as this framework is still under development. 
- More QueryCases and Expected Results are needed.

Running:
$ gradle run -PappArgs="['-bt','CONFORMANCE','-st','MEMORY']"